Dry run check done in the save attachment on disk method

+++ mailbag/helper.py
@@ -78,14 +78,15 @@ def moveWithDirectoryStructure(dry_run, mainPath, mailbag_name, input, emailFold
 def saveAttachments(part):
     return (part.get_filename(),part.get_payload(decode=True))
 
-def saveAttachmentOnDisk(attachments_dir,message):
+def saveAttachmentOnDisk(dry_run,attachments_dir,message):
     
     message_attachments_dir = os.path.join(attachments_dir,str(message.Mailbag_Message_ID))
     os.mkdir(message_attachments_dir)
     for i in range(message.AttachmentNum):
         log.debug('Saving Attachment:'+str(message.AttachmentNames[i]))
-        attachment_path = os.path.join(message_attachments_dir,message.AttachmentNames[i])
-        f = open(attachment_path, "wb")
-        f.write(message.AttachmentFiles[i])
-        f.close()
+        if not dry_run:
+            attachment_path = os.path.join(message_attachments_dir,message.AttachmentNames[i])
+            f = open(attachment_path, "wb")
+            f.write(message.AttachmentFiles[i])
+            f.close()
